Reorganize test according to src directory
